requirements.txt
pymupdf
langdetect
sentence-transformers
faiss-cpu
numpy
